add game api in blazor app
add api of add, get all game, delete in blazor. get all game is perfectly fine. but add and delete did not end well as add is trigger and show 200 but is not added and delete button not respond. that i stuck on api that need to use button to trigger is not working, but if i do it raw way is working